# Optimizations to recordPrevalence and related functions