--- a/src/migrations.js
+++ b/src/migrations.js
@@ -65,8 +65,7 @@ exports.Migrations= {
     // Import snitch map 
     _.each(originFrequency, function(fpDomains, trackingDomain){
       _.each(fpDomains, function(v, firstParty){
-        badger.heuristicBlocking.recordPrevalence(
-          trackingDomain,
+        badger.heuristicBlocking.updateTrackerPrevalence(
           trackingDomain,
           firstParty
         );