--- a/src/webrequest.js
+++ b/src/webrequest.js
@@ -408,8 +408,8 @@ function recordFingerprinting(tabId, msg) {
           log(script_host, 'caught fingerprinting on', document_host);
 
           // Mark this is a strike
-          badger.heuristicBlocking.recordPrevalence(
-            script_host, script_origin, window.getBaseDomain(document_host));
+          badger.heuristicBlocking.updateTrackerPrevalence(
+            script_host, window.getBaseDomain(document_host));
         }
       }
       // This is a canvas write