Here's a first stab at the refactoring of the `recordPrevalence` call path. It turns out most of the code from `heuristicBlockingAccounting` (didn't modify the name here, but we should at some point) was superfluous when operating under the assumption that anything in `snitch_map` from an import was already confirmed to have been tracking (as are any calls from `recordFingerprinting`).

Above is the "Partial" optimization. The "Full" optimization includes the other optimization present in this pull, namely the removal of two calls to `setupHeuristicAction` for `tracker_fqdn` and `tracker_origin` in `heuristicblocking.js` line 204:

```javascript
if(firstParties.indexOf(page_origin) === -1){
  firstParties.push(page_origin);
  snitch_map.setItem(tracker_origin, firstParties);
} 
```


Benchmarks are as follows, with "Runtime" being the time that the Privacy Badger extension runs at 100% CPU. Set size is 600 tracker domains with 8300 first parties:

| Version    | Test        | Runtime   |
| :------------- |:------------- |:----- |
| Current      | Fresh import | 3m30s |
| Opt(Partial) | Fresh import   |    50s |
| Opt(Full) | Fresh import   |    0s |
| Current      | Re-import      |   3m30s |
| Opt(Partial) | Re-import      |  0s |
| Opt(Full) | Re-import   |    0s |


Still, the removal of those two lines from `heuristicblocking.js` causes tests to fail and I'm not sure I entirely understand why. They appear to be responsible for the last major vestige of computational overhead associated with importing, so any help in better explaining why we need these calls would be super helpful!