@alexristich 2017-01-18
Seems to actually be quite fast in this form. 100% CPU time looks to have been at most a second or two. I'm having trouble running the tests on my local machine again - I'm getting the error `Xvfb did not start`. (@cooperq Do you remember how I managed to fix this last time? I neglected to write it down and am regretting it right about now!)