@ghostwords 2017-01-18
To run QUnit tests, just [visit the tests.html page in your browser](https://github.com/EFForg/privacybadger/blob/master/README.md#testing).