@ghostwords 2017-01-18
Here's a first pass at a debounced-per-argument function:
```javascript
var debouncedFunc = (function () {
  var funcs = {}; // preserved by the closure
  return function (a) {
    if (!funcs.hasOwnProperty(a)) {
      funcs[a] = _.debounce(function () {
        FUNCTION_BEING_DEBOUNCED(a);
      });
    }
    funcs[a]();
  };
}());
```