@alexristich 2017-01-18
@ghostwords Thanks! I can't run any Selenium tests, but I probably should've looked at the README sooner to figure out how to run the QUnit tests :)

Just pushed a change to one of the tests. It was failing because we were manually adding it to `action_map` without a `heuristicAction`. I don't know off the top of my head if there are cases where items are being added to `action_map` without a heuristicAction being set. 

Perhaps another change needs to be made too, though - I'm not convinced the current logic is correct. Specifically, if a domain has `heuristicAction: "notracking"`, it would be passed in to `_recordPrevalence()` but would retain its `notracking` title until it hit the threshold. Does it make sense to also check to see if `heuristicAction === constants.NO_TRACKING` then change it to `constants.ALLOW`? We know that once a domain has reached `_recordPrevalence()` we are already assured that it is tracking so perhaps its a safe bet?