@alexristich 2017-01-21
By the way, the test that is failing here seems to be failing on the current master branch as well. I'm not sure what is causing it, but at first glance it doesn't appear that anything in this PR is breaking any tests that are currently passing on master.