@ghostwords 2017-01-23
@cooperq I don't think we should include this in today's release. I would like to give it another round of consideration tomorrow.