@ghostwords 2017-01-23
OK, I'm done for now, please take a look at my changes. I'll re-review tomorrow with fresh eyes and sign off then hopefully.