@alexristich 2017-01-23
On first pass, this definitely looks better - also, good call on adding a link to this discussion for future reference.